뭐든 추가 했겠찌

+++ src/main/java/com/green/vrink/community/controller/FreeBoardReplyRestController.java
@@ -39,6 +39,9 @@ public class FreeBoardReplyRestController {
             @PathVariable(name = "commu-id")
             Integer commuId
     ) {
+        if (commuId == null){
+            return ResponseEntity.badRequest().build();
+        }
         freeBoardReplyService.delete(commuId);
         return ResponseEntity.ok().build();
     }
